refactor: make less ffi-centric
